# openapi.yaml
openapi: 3.1.0
info:
  title: Go Agent Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8081/api/v1
paths:
  /auth/login:
    post:
      summary: Login with seeded admin credentials or future user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authenticated
  /me:
    get:
      summary: Get current user and workspaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
  /agents:
    get:
      summary: List agents in a workspace
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: workspace_id
          schema:
            type: string
      responses:
        "200":
          description: Agent list
    post:
      summary: Create agent
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Agent created
  /tools:
    get:
      summary: List tools in a workspace
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tool list
    post:
      summary: Create tool
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Tool created
  /tasks:
    post:
      summary: Execute a task
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Task accepted
  /tasks/{taskId}:
    get:
      summary: Get task detail and steps
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
  /approvals/{approvalId}/approve:
    post:
      summary: Approve a pending approval
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: approvalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval updated
  /approvals/{approvalId}/reject:
    post:
      summary: Reject a pending approval
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: approvalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer